requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
groq==0.13.1
chromadb==0.5.23
sentence-transformers==3.3.1
pyjwt==2.10.1
bcrypt==4.2.1
slowapi==0.1.9
jsonlines==4.0.0
tabulate==0.9.0
pydantic==2.10.3
python-multipart==0.0.19
